+++ src/EnrolmentWorkflow/Rule/Fields/Field/FieldFormGUI.php
@@ -3,6 +3,7 @@
 namespace srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\Fields\Field;
 
 use ilTextInputGUI;
+use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\RulesGUI;
 
 /**
  * Trait FieldFormGUI
@@ -22,7 +23,8 @@ trait FieldFormGUI
         return [
             "field" => [
                 self::PROPERTY_CLASS    => ilTextInputGUI::class,
-                self::PROPERTY_REQUIRED => true
+                self::PROPERTY_REQUIRED => true,
+                "setTitle"              => self::plugin()->translate("field", RulesGUI::LANG_MODULE)
             ]
         ];
     }