+++ src/EnrolmentWorkflow/Rule/Fields/Position/PositionConstants.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\Fields\Position;
+
+/**
+ * Class PositionConstants
+ *
+ * @package srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\Fields\Position
+ *
+ * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
+ */
+final class PositionConstants
+{
+
+    const POSITION_ALL = 0;
+
+
+    /**
+     * OperatorConstants constructor
+     */
+    private function __construct()
+    {
+
+    }
+}