+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [x]
+## [3.5.0]
 - Rule 'Current user is assigned as responsible user'
 - Filter recursive and UDF in action 'Assign responsible users'
 - Fixes